Crate mod_redis

Modules

test

Structs

ClientManager

Cmd

Represents redis commands.

RedisConnKey

RedisConnection

RedisError

Represents a redis error. For the most part you should be using the Error trait to interact with this rather than the actual struct.

Script

Represents a lua script.

ScriptInvocation

Represents a prepared script call.

Enums

ClientWrapper

ConnectionWrapper

NodeSpec

RedisValue

Internal low-level redis value enum.

Traits

FromRedisValue

This trait is used to convert a redis value into a more appropriate type. While a redis Value can represent any response that comes back from the redis server, usually you want to map this into something that works better in rust. For instance you might want to convert the return value into a String or an integer.

Functions

build_cmd

cmd

Shortcut function to creating a command with a single argument.

register