mod_redis

Enum RedisValue

pub enum RedisValue {

    Nil,
    Int(i64),
    BulkString(Vec<u8>),
    Array(Vec<Value>),
    SimpleString(String),
    Okay,
    Map(Vec<(Value, Value)>),
    Attribute {
        data: Box<Value>,
        attributes: Vec<(Value, Value)>,
    },
    Set(Vec<Value>),
    Double(f64),
    Boolean(bool),
    VerbatimString {
        format: VerbatimFormat,
        text: String,
    },
    BigNumber(BigInt),
    Push {
        kind: PushKind,
        data: Vec<Value>,
    },
    ServerError(ServerError),

}

Internal low-level redis value enum.

Variants

Nil

A nil response from the server.

Int(i64)

An integer response. Note that there are a few situations in which redis actually returns a string for an integer which is why this library generally treats integers and strings the same for all numeric responses.

BulkString(Vec<u8>)

An arbitrary binary data, usually represents a binary-safe string.

Array(Vec<Value>)

A response containing an array with more data. This is generally used by redis to express nested structures.

SimpleString(String)

A simple string response, without line breaks and not binary safe.

Okay

A status response which represents the string โOKโ.

Map(Vec<(Value, Value)>)

Unordered key,value list from the server. Use as_map_iter function.

Attribute

Attribute value from the server. Client will give data instead of whole Attribute type.

Fields

data: Box<Value>

Data that attributes belong to.

attributes: Vec<(Value, Value)>

Key,Value list of attributes.

Set(Vec<Value>)

Unordered set value from the server.

Double(f64)

A floating number response from the server.

Boolean(bool)

A boolean response from the server.

VerbatimString

First String is format and other is the string

Fields

format: VerbatimFormat

Textโs format type

text: String

Remaining string check format before using!

BigNumber(BigInt)

Very large number that out of the range of the signed 64 bit numbers

Push

Push data from the server.

Fields

kind: PushKind

Push Kind

data: Vec<Value>

Remaining data from push message

ServerError(ServerError)

Represents an error message from the server

Implementations

impl Value

Values are generally not used directly unless you are using the more low level functionality in the library. For the most part this is hidden with the help of the FromRedisValue trait.

While on the redis protocol there is an error type this is already separated at an early point so the value only holds the remaining types.

pub fn looks_like_cursor(&self) -> bool

Checks if the return value looks like it fulfils the cursor protocol. That means the result is an array item of length two with the first one being a cursor and the second an array response.

pub fn as_sequence(&self) -> Option<&[Value]>

Returns an &[Value] if self is compatible with a sequence type

pub fn into_sequence(self) -> Result<Vec<Value>, Value>

Returns a Vec<Value> if self is compatible with a sequence type, otherwise returns Err(self).

pub fn as_map_iter(&self) -> Option<MapIter<'_>>

Returns an iterator of (&Value, &Value) if self is compatible with a map type

pub fn into_map_iter(self) -> Result<OwnedMapIter, Value>

Returns an iterator of (Value, Value) if self is compatible with a map type. If not, returns Err(self).

pub fn extract_error(self) -> Result<Value, RedisError>

If value contains a server error, return it as an Err. Otherwise wrap the value in Ok.