Struct mod_redis::Cmd

pub struct Cmd { /* private fields */ }

Represents redis commands.

Implementations

impl Cmd

A command acts as a builder interface to creating encoded redis requests. This allows you to easiy assemble a packed command by chaining arguments together.

Basic example:

redis::Cmd::new().arg("SET").arg("my_key").arg(42);

There is also a helper function called cmd which makes it a tiny bit shorter:

redis::cmd("SET").arg("my_key").arg(42);

Because Rust currently does not have an ideal system for lifetimes of temporaries, sometimes you need to hold on to the initially generated command:

let mut cmd = redis::cmd("SMEMBERS");
let mut iter : redis::Iter<i32> = cmd.arg("my_set").clone().iter(&mut con).unwrap();

pub fn new() -> Cmd

Creates a new empty command.

pub fn arg<T>(&mut self, arg: T) -> &mut Cmd
where T: ToRedisArgs,

Appends an argument to the command. The argument passed must be a type that implements ToRedisArgs. Most primitive types as well as vectors of primitive types implement it.

For instance all of the following are valid:

redis::cmd("SET").arg(&["my_key", "my_value"]);
redis::cmd("SET").arg("my_key").arg(42);
redis::cmd("SET").arg("my_key").arg(b"my_value");

pub fn cursor_arg(&mut self, cursor: u64) -> &mut Cmd

Works similar to arg but adds a cursor argument. This is always an integer and also flips the command implementation to support a different mode for the iterators where the iterator will ask for another batch of items when the local data is exhausted.

let mut cmd = redis::cmd("SSCAN");
let mut iter : redis::Iter<isize> =
    cmd.arg("my_set").cursor_arg(0).clone().iter(&mut con).unwrap();
for x in iter {
    // do something with the item
}

pub fn get_packed_command(&self) -> Vec<u8>

Returns the packed command as a byte vector.

pub fn in_scan_mode(&self) -> bool

Returns true if the command is in scan mode.

pub fn query<T>(&self, con: &mut dyn ConnectionLike) -> Result<T, RedisError>
where T: FromRedisValue,

Sends the command as query to the connection and converts the result to the target redis value. This is the general way how you can retrieve data.

pub fn iter<T>( self, con: &mut dyn ConnectionLike ) -> Result<Iter<'_, T>, RedisError>
where T: FromRedisValue,

Similar to query() but returns an iterator over the items of the bulk result or iterator. In normal mode this is not in any way more efficient than just querying into a Vec<T> as it’s internally implemented as buffering into a vector. This however is useful when cursor_arg was used in which case the iterator will query for more items until the server side cursor is exhausted.

This is useful for commands such as SSCAN, SCAN and others.

One speciality of this function is that it will check if the response looks like a cursor or not and always just looks at the payload. This way you can use the function the same for responses in the format of KEYS (just a list) as well as SSCAN (which returns a tuple of cursor and list).

pub fn execute(&self, con: &mut dyn ConnectionLike)

This is a shortcut to query() that does not return a value and will fail the task if the query fails because of an error. This is mainly useful in examples and for simple commands like setting keys.

This is equivalent to a call of query like this:

let _ : () = redis::cmd("PING").query(&mut con).unwrap();

pub fn args_iter( &self ) -> impl Iterator<Item = Arg<&[u8]>> + Clone + ExactSizeIterator

Returns an iterator over the arguments in this command (including the command name itself)

impl Cmd

pub fn get<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Get the value of a key. If key is a vec this becomes an MGET.

pub fn mget<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Get values of keys

pub fn keys<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Gets all keys matching pattern

pub fn set<K, V, 'a>(key: K, value: V) -> Cmd
where K: ToRedisArgs, V: ToRedisArgs,

Set the string value of a key.

pub fn set_multiple<K, V, 'a>(items: &'a [(K, V)]) -> Cmd
where K: ToRedisArgs, V: ToRedisArgs,

Sets multiple keys to their values.

pub fn set_ex<K, V, 'a>(key: K, value: V, seconds: usize) -> Cmd
where K: ToRedisArgs, V: ToRedisArgs,

Set the value and expiration of a key.

pub fn pset_ex<K, V, 'a>(key: K, value: V, milliseconds: usize) -> Cmd
where K: ToRedisArgs, V: ToRedisArgs,

Set the value and expiration in milliseconds of a key.

pub fn set_nx<K, V, 'a>(key: K, value: V) -> Cmd
where K: ToRedisArgs, V: ToRedisArgs,

Set the value of a key, only if the key does not exist

pub fn mset_nx<K, V, 'a>(items: &'a [(K, V)]) -> Cmd
where K: ToRedisArgs, V: ToRedisArgs,

Sets multiple keys to their values failing if at least one already exists.

pub fn getset<K, V, 'a>(key: K, value: V) -> Cmd
where K: ToRedisArgs, V: ToRedisArgs,

Set the string value of a key and return its old value.

pub fn getrange<K, 'a>(key: K, from: isize, to: isize) -> Cmd
where K: ToRedisArgs,

Get a range of bytes/substring from the value of a key. Negative values provide an offset from the end of the value.

pub fn setrange<K, V, 'a>(key: K, offset: isize, value: V) -> Cmd
where K: ToRedisArgs, V: ToRedisArgs,

Overwrite the part of the value stored in key at the specified offset.

pub fn del<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Delete one or more keys.

pub fn exists<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Determine if a key exists.

pub fn expire<K, 'a>(key: K, seconds: usize) -> Cmd
where K: ToRedisArgs,

Set a key’s time to live in seconds.

pub fn expire_at<K, 'a>(key: K, ts: usize) -> Cmd
where K: ToRedisArgs,

Set the expiration for a key as a UNIX timestamp.

pub fn pexpire<K, 'a>(key: K, ms: usize) -> Cmd
where K: ToRedisArgs,

Set a key’s time to live in milliseconds.

pub fn pexpire_at<K, 'a>(key: K, ts: usize) -> Cmd
where K: ToRedisArgs,

Set the expiration for a key as a UNIX timestamp in milliseconds.

pub fn persist<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Remove the expiration from a key.

pub fn ttl<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Get the expiration time of a key.

pub fn pttl<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Get the expiration time of a key in milliseconds.

pub fn get_ex<K, 'a>(key: K, expire_at: Expiry) -> Cmd
where K: ToRedisArgs,

Get the value of a key and set expiration

pub fn get_del<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Get the value of a key and delete it

pub fn rename<K, 'a>(key: K, new_key: K) -> Cmd
where K: ToRedisArgs,

Rename a key.

pub fn rename_nx<K, 'a>(key: K, new_key: K) -> Cmd
where K: ToRedisArgs,

Rename a key, only if the new key does not exist.

pub fn unlink<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Unlink one or more keys.

pub fn append<K, V, 'a>(key: K, value: V) -> Cmd
where K: ToRedisArgs, V: ToRedisArgs,

Append a value to a key.

pub fn incr<K, V, 'a>(key: K, delta: V) -> Cmd
where K: ToRedisArgs, V: ToRedisArgs,

Increment the numeric value of a key by the given amount. This issues a INCRBY or INCRBYFLOAT depending on the type.

pub fn decr<K, V, 'a>(key: K, delta: V) -> Cmd
where K: ToRedisArgs, V: ToRedisArgs,

Decrement the numeric value of a key by the given amount.

pub fn setbit<K, 'a>(key: K, offset: usize, value: bool) -> Cmd
where K: ToRedisArgs,

Sets or clears the bit at offset in the string value stored at key.

pub fn getbit<K, 'a>(key: K, offset: usize) -> Cmd
where K: ToRedisArgs,

Returns the bit value at offset in the string value stored at key.

pub fn bitcount<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Count set bits in a string.

pub fn bitcount_range<K, 'a>(key: K, start: usize, end: usize) -> Cmd
where K: ToRedisArgs,

Count set bits in a string in a range.

pub fn bit_and<K, 'a>(dstkey: K, srckeys: K) -> Cmd
where K: ToRedisArgs,

Perform a bitwise AND between multiple keys (containing string values) and store the result in the destination key.

pub fn bit_or<K, 'a>(dstkey: K, srckeys: K) -> Cmd
where K: ToRedisArgs,

Perform a bitwise OR between multiple keys (containing string values) and store the result in the destination key.

pub fn bit_xor<K, 'a>(dstkey: K, srckeys: K) -> Cmd
where K: ToRedisArgs,

Perform a bitwise XOR between multiple keys (containing string values) and store the result in the destination key.

pub fn bit_not<K, 'a>(dstkey: K, srckey: K) -> Cmd
where K: ToRedisArgs,

Perform a bitwise NOT of the key (containing string values) and store the result in the destination key.

pub fn strlen<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Get the length of the value stored in a key.

pub fn hget<K, F, 'a>(key: K, field: F) -> Cmd
where K: ToRedisArgs, F: ToRedisArgs,

Gets a single (or multiple) fields from a hash.

pub fn hdel<K, F, 'a>(key: K, field: F) -> Cmd
where K: ToRedisArgs, F: ToRedisArgs,

Deletes a single (or multiple) fields from a hash.

pub fn hset<K, F, V, 'a>(key: K, field: F, value: V) -> Cmd
where K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs,

Sets a single field in a hash.

pub fn hset_nx<K, F, V, 'a>(key: K, field: F, value: V) -> Cmd
where K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs,

Sets a single field in a hash if it does not exist.

pub fn hset_multiple<K, F, V, 'a>(key: K, items: &'a [(F, V)]) -> Cmd
where K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs,

Sets a multiple fields in a hash.

pub fn hincr<K, F, D, 'a>(key: K, field: F, delta: D) -> Cmd
where K: ToRedisArgs, F: ToRedisArgs, D: ToRedisArgs,

Increments a value.

pub fn hexists<K, F, 'a>(key: K, field: F) -> Cmd
where K: ToRedisArgs, F: ToRedisArgs,

Checks if a field in a hash exists.

pub fn hkeys<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Gets all the keys in a hash.

pub fn hvals<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Gets all the values in a hash.

pub fn hgetall<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Gets all the fields and values in a hash.

pub fn hlen<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Gets the length of a hash.

pub fn blmove<K, 'a>( srckey: K, dstkey: K, src_dir: Direction, dst_dir: Direction, timeout: usize ) -> Cmd
where K: ToRedisArgs,

Pop an element from a list, push it to another list and return it; or block until one is available

pub fn blmpop<K, 'a>( timeout: usize, numkeys: usize, key: K, dir: Direction, count: usize ) -> Cmd
where K: ToRedisArgs,

Pops count elements from the first non-empty list key from the list of provided key names; or blocks until one is available.

pub fn blpop<K, 'a>(key: K, timeout: usize) -> Cmd
where K: ToRedisArgs,

Remove and get the first element in a list, or block until one is available.

pub fn brpop<K, 'a>(key: K, timeout: usize) -> Cmd
where K: ToRedisArgs,

Remove and get the last element in a list, or block until one is available.

pub fn brpoplpush<K, 'a>(srckey: K, dstkey: K, timeout: usize) -> Cmd
where K: ToRedisArgs,

Pop a value from a list, push it to another list and return it; or block until one is available.

pub fn lindex<K, 'a>(key: K, index: isize) -> Cmd
where K: ToRedisArgs,

Get an element from a list by its index.

pub fn linsert_before<K, P, V, 'a>(key: K, pivot: P, value: V) -> Cmd
where K: ToRedisArgs, P: ToRedisArgs, V: ToRedisArgs,

Insert an element before another element in a list.

pub fn linsert_after<K, P, V, 'a>(key: K, pivot: P, value: V) -> Cmd
where K: ToRedisArgs, P: ToRedisArgs, V: ToRedisArgs,

Insert an element after another element in a list.

pub fn llen<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Returns the length of the list stored at key.

pub fn lmove<K, 'a>( srckey: K, dstkey: K, src_dir: Direction, dst_dir: Direction ) -> Cmd
where K: ToRedisArgs,

Pop an element a list, push it to another list and return it

pub fn lmpop<K, 'a>(numkeys: usize, key: K, dir: Direction, count: usize) -> Cmd
where K: ToRedisArgs,

Pops count elements from the first non-empty list key from the list of provided key names.

pub fn lpop<K, 'a>(key: K, count: Option<NonZeroUsize>) -> Cmd
where K: ToRedisArgs,

Removes and returns the up to count first elements of the list stored at key.

If count is not specified, then defaults to first element.

pub fn lpos<K, V, 'a>(key: K, value: V, options: LposOptions) -> Cmd
where K: ToRedisArgs, V: ToRedisArgs,

Returns the index of the first matching value of the list stored at key.

pub fn lpush<K, V, 'a>(key: K, value: V) -> Cmd
where K: ToRedisArgs, V: ToRedisArgs,

Insert all the specified values at the head of the list stored at key.

pub fn lpush_exists<K, V, 'a>(key: K, value: V) -> Cmd
where K: ToRedisArgs, V: ToRedisArgs,

Inserts a value at the head of the list stored at key, only if key already exists and holds a list.

pub fn lrange<K, 'a>(key: K, start: isize, stop: isize) -> Cmd
where K: ToRedisArgs,

Returns the specified elements of the list stored at key.

pub fn lrem<K, V, 'a>(key: K, count: isize, value: V) -> Cmd
where K: ToRedisArgs, V: ToRedisArgs,

Removes the first count occurrences of elements equal to value from the list stored at key.

pub fn ltrim<K, 'a>(key: K, start: isize, stop: isize) -> Cmd
where K: ToRedisArgs,

Trim an existing list so that it will contain only the specified range of elements specified.

pub fn lset<K, V, 'a>(key: K, index: isize, value: V) -> Cmd
where K: ToRedisArgs, V: ToRedisArgs,

Sets the list element at index to value

pub fn rpop<K, 'a>(key: K, count: Option<NonZeroUsize>) -> Cmd
where K: ToRedisArgs,

Removes and returns the up to count last elements of the list stored at key

If count is not specified, then defaults to last element.

pub fn rpoplpush<K, 'a>(key: K, dstkey: K) -> Cmd
where K: ToRedisArgs,

Pop a value from a list, push it to another list and return it.

pub fn rpush<K, V, 'a>(key: K, value: V) -> Cmd
where K: ToRedisArgs, V: ToRedisArgs,

Insert all the specified values at the tail of the list stored at key.

pub fn rpush_exists<K, V, 'a>(key: K, value: V) -> Cmd
where K: ToRedisArgs, V: ToRedisArgs,

Inserts value at the tail of the list stored at key, only if key already exists and holds a list.

pub fn sadd<K, M, 'a>(key: K, member: M) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs,

Add one or more members to a set.

pub fn scard<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Get the number of members in a set.

pub fn sdiff<K, 'a>(keys: K) -> Cmd
where K: ToRedisArgs,

Subtract multiple sets.

pub fn sdiffstore<K, 'a>(dstkey: K, keys: K) -> Cmd
where K: ToRedisArgs,

Subtract multiple sets and store the resulting set in a key.

pub fn sinter<K, 'a>(keys: K) -> Cmd
where K: ToRedisArgs,

Intersect multiple sets.

pub fn sinterstore<K, 'a>(dstkey: K, keys: K) -> Cmd
where K: ToRedisArgs,

Intersect multiple sets and store the resulting set in a key.

pub fn sismember<K, M, 'a>(key: K, member: M) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs,

Determine if a given value is a member of a set.

pub fn smembers<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Get all the members in a set.

pub fn smove<K, M, 'a>(srckey: K, dstkey: K, member: M) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs,

Move a member from one set to another.

pub fn spop<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Remove and return a random member from a set.

pub fn srandmember<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Get one random member from a set.

pub fn srandmember_multiple<K, 'a>(key: K, count: usize) -> Cmd
where K: ToRedisArgs,

Get multiple random members from a set.

pub fn srem<K, M, 'a>(key: K, member: M) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs,

Remove one or more members from a set.

pub fn sunion<K, 'a>(keys: K) -> Cmd
where K: ToRedisArgs,

Add multiple sets.

pub fn sunionstore<K, 'a>(dstkey: K, keys: K) -> Cmd
where K: ToRedisArgs,

Add multiple sets and store the resulting set in a key.

pub fn zadd<K, S, M, 'a>(key: K, member: M, score: S) -> Cmd
where K: ToRedisArgs, S: ToRedisArgs, M: ToRedisArgs,

Add one member to a sorted set, or update its score if it already exists.

pub fn zadd_multiple<K, S, M, 'a>(key: K, items: &'a [(S, M)]) -> Cmd
where K: ToRedisArgs, S: ToRedisArgs, M: ToRedisArgs,

Add multiple members to a sorted set, or update its score if it already exists.

pub fn zcard<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Get the number of members in a sorted set.

pub fn zcount<K, M, MM, 'a>(key: K, min: M, max: MM) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs,

Count the members in a sorted set with scores within the given values.

pub fn zincr<K, M, D, 'a>(key: K, member: M, delta: D) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs, D: ToRedisArgs,

Increments the member in a sorted set at key by delta. If the member does not exist, it is added with delta as its score.

pub fn zinterstore<K, 'a>(dstkey: K, keys: &'a [K]) -> Cmd
where K: ToRedisArgs,

Intersect multiple sorted sets and store the resulting sorted set in a new key using SUM as aggregation function.

pub fn zinterstore_min<K, 'a>(dstkey: K, keys: &'a [K]) -> Cmd
where K: ToRedisArgs,

Intersect multiple sorted sets and store the resulting sorted set in a new key using MIN as aggregation function.

pub fn zinterstore_max<K, 'a>(dstkey: K, keys: &'a [K]) -> Cmd
where K: ToRedisArgs,

Intersect multiple sorted sets and store the resulting sorted set in a new key using MAX as aggregation function.

pub fn zinterstore_weights<K, W, 'a>(dstkey: K, keys: &'a [(K, W)]) -> Cmd
where K: ToRedisArgs, W: ToRedisArgs,

[Commands::zinterstore], but with the ability to specify a multiplication factor for each sorted set by pairing one with each key in a tuple.

pub fn zinterstore_min_weights<K, W, 'a>(dstkey: K, keys: &'a [(K, W)]) -> Cmd
where K: ToRedisArgs, W: ToRedisArgs,

[Commands::zinterstore_min], but with the ability to specify a multiplication factor for each sorted set by pairing one with each key in a tuple.

pub fn zinterstore_max_weights<K, W, 'a>(dstkey: K, keys: &'a [(K, W)]) -> Cmd
where K: ToRedisArgs, W: ToRedisArgs,

[Commands::zinterstore_max], but with the ability to specify a multiplication factor for each sorted set by pairing one with each key in a tuple.

pub fn zlexcount<K, L, 'a>(key: K, min: L, max: L) -> Cmd
where K: ToRedisArgs, L: ToRedisArgs,

Count the number of members in a sorted set between a given lexicographical range.

pub fn zpopmax<K, 'a>(key: K, count: isize) -> Cmd
where K: ToRedisArgs,

Removes and returns up to count members with the highest scores in a sorted set

pub fn zpopmin<K, 'a>(key: K, count: isize) -> Cmd
where K: ToRedisArgs,

Removes and returns up to count members with the lowest scores in a sorted set

pub fn zmpop_max<K, 'a>(keys: &'a [K], count: isize) -> Cmd
where K: ToRedisArgs,

Removes and returns up to count members with the highest scores, from the first non-empty sorted set in the provided list of key names.

pub fn zmpop_min<K, 'a>(keys: &'a [K], count: isize) -> Cmd
where K: ToRedisArgs,

Removes and returns up to count members with the lowest scores, from the first non-empty sorted set in the provided list of key names.

pub fn zrandmember<K, 'a>(key: K, count: Option<isize>) -> Cmd
where K: ToRedisArgs,

Return up to count random members in a sorted set (or 1 if count == None)

pub fn zrandmember_withscores<K, 'a>(key: K, count: isize) -> Cmd
where K: ToRedisArgs,

Return up to count random members in a sorted set with scores

pub fn zrange<K, 'a>(key: K, start: isize, stop: isize) -> Cmd
where K: ToRedisArgs,

Return a range of members in a sorted set, by index

pub fn zrange_withscores<K, 'a>(key: K, start: isize, stop: isize) -> Cmd
where K: ToRedisArgs,

Return a range of members in a sorted set, by index with scores.

pub fn zrangebylex<K, M, MM, 'a>(key: K, min: M, max: MM) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs,

Return a range of members in a sorted set, by lexicographical range.

pub fn zrangebylex_limit<K, M, MM, 'a>( key: K, min: M, max: MM, offset: isize, count: isize ) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs,

Return a range of members in a sorted set, by lexicographical range with offset and limit.

pub fn zrevrangebylex<K, MM, M, 'a>(key: K, max: MM, min: M) -> Cmd
where K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs,

Return a range of members in a sorted set, by lexicographical range.

pub fn zrevrangebylex_limit<K, MM, M, 'a>( key: K, max: MM, min: M, offset: isize, count: isize ) -> Cmd
where K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs,

Return a range of members in a sorted set, by lexicographical range with offset and limit.

pub fn zrangebyscore<K, M, MM, 'a>(key: K, min: M, max: MM) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs,

Return a range of members in a sorted set, by score.

pub fn zrangebyscore_withscores<K, M, MM, 'a>(key: K, min: M, max: MM) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs,

Return a range of members in a sorted set, by score with scores.

pub fn zrangebyscore_limit<K, M, MM, 'a>( key: K, min: M, max: MM, offset: isize, count: isize ) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs,

Return a range of members in a sorted set, by score with limit.

pub fn zrangebyscore_limit_withscores<K, M, MM, 'a>( key: K, min: M, max: MM, offset: isize, count: isize ) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs,

Return a range of members in a sorted set, by score with limit with scores.

pub fn zrank<K, M, 'a>(key: K, member: M) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs,

Determine the index of a member in a sorted set.

pub fn zrem<K, M, 'a>(key: K, members: M) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs,

Remove one or more members from a sorted set.

pub fn zrembylex<K, M, MM, 'a>(key: K, min: M, max: MM) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs,

Remove all members in a sorted set between the given lexicographical range.

pub fn zremrangebyrank<K, 'a>(key: K, start: isize, stop: isize) -> Cmd
where K: ToRedisArgs,

Remove all members in a sorted set within the given indexes.

pub fn zrembyscore<K, M, MM, 'a>(key: K, min: M, max: MM) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs,

Remove all members in a sorted set within the given scores.

pub fn zrevrange<K, 'a>(key: K, start: isize, stop: isize) -> Cmd
where K: ToRedisArgs,

Return a range of members in a sorted set, by index, with scores ordered from high to low.

pub fn zrevrange_withscores<K, 'a>(key: K, start: isize, stop: isize) -> Cmd
where K: ToRedisArgs,

Return a range of members in a sorted set, by index, with scores ordered from high to low.

pub fn zrevrangebyscore<K, MM, M, 'a>(key: K, max: MM, min: M) -> Cmd
where K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs,

Return a range of members in a sorted set, by score.

pub fn zrevrangebyscore_withscores<K, MM, M, 'a>(key: K, max: MM, min: M) -> Cmd
where K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs,

Return a range of members in a sorted set, by score with scores.

pub fn zrevrangebyscore_limit<K, MM, M, 'a>( key: K, max: MM, min: M, offset: isize, count: isize ) -> Cmd
where K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs,

Return a range of members in a sorted set, by score with limit.

pub fn zrevrangebyscore_limit_withscores<K, MM, M, 'a>( key: K, max: MM, min: M, offset: isize, count: isize ) -> Cmd
where K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs,

Return a range of members in a sorted set, by score with limit with scores.

pub fn zrevrank<K, M, 'a>(key: K, member: M) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs,

Determine the index of a member in a sorted set, with scores ordered from high to low.

pub fn zscore<K, M, 'a>(key: K, member: M) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs,

Get the score associated with the given member in a sorted set.

pub fn zscore_multiple<K, M, 'a>(key: K, members: &'a [M]) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs,

Get the scores associated with multiple members in a sorted set.

pub fn zunionstore<K, 'a>(dstkey: K, keys: &'a [K]) -> Cmd
where K: ToRedisArgs,

Unions multiple sorted sets and store the resulting sorted set in a new key using SUM as aggregation function.

pub fn zunionstore_min<K, 'a>(dstkey: K, keys: &'a [K]) -> Cmd
where K: ToRedisArgs,

Unions multiple sorted sets and store the resulting sorted set in a new key using MIN as aggregation function.

pub fn zunionstore_max<K, 'a>(dstkey: K, keys: &'a [K]) -> Cmd
where K: ToRedisArgs,

Unions multiple sorted sets and store the resulting sorted set in a new key using MAX as aggregation function.

pub fn zunionstore_weights<K, W, 'a>(dstkey: K, keys: &'a [(K, W)]) -> Cmd
where K: ToRedisArgs, W: ToRedisArgs,

[Commands::zunionstore], but with the ability to specify a multiplication factor for each sorted set by pairing one with each key in a tuple.

pub fn zunionstore_min_weights<K, W, 'a>(dstkey: K, keys: &'a [(K, W)]) -> Cmd
where K: ToRedisArgs, W: ToRedisArgs,

[Commands::zunionstore_min], but with the ability to specify a multiplication factor for each sorted set by pairing one with each key in a tuple.

pub fn zunionstore_max_weights<K, W, 'a>(dstkey: K, keys: &'a [(K, W)]) -> Cmd
where K: ToRedisArgs, W: ToRedisArgs,

[Commands::zunionstore_max], but with the ability to specify a multiplication factor for each sorted set by pairing one with each key in a tuple.

pub fn pfadd<K, E, 'a>(key: K, element: E) -> Cmd
where K: ToRedisArgs, E: ToRedisArgs,

Adds the specified elements to the specified HyperLogLog.

pub fn pfcount<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Return the approximated cardinality of the set(s) observed by the HyperLogLog at key(s).

pub fn pfmerge<K, 'a>(dstkey: K, srckeys: K) -> Cmd
where K: ToRedisArgs,

Merge N different HyperLogLogs into a single one.

pub fn publish<K, E, 'a>(channel: K, message: E) -> Cmd
where K: ToRedisArgs, E: ToRedisArgs,

Posts a message to the given channel.

pub fn object_encoding<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Returns the encoding of a key.

pub fn object_idletime<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Returns the time in seconds since the last access of a key.

pub fn object_freq<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Returns the logarithmic access frequency counter of a key.

pub fn object_refcount<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Returns the reference count of a key.

pub fn acl_load<'a>() -> Cmd

When Redis is configured to use an ACL file (with the aclfile configuration option), this command will reload the ACLs from the file, replacing all the current ACL rules with the ones defined in the file.

pub fn acl_save<'a>() -> Cmd

When Redis is configured to use an ACL file (with the aclfile configuration option), this command will save the currently defined ACLs from the server memory to the ACL file.

pub fn acl_list<'a>() -> Cmd

Shows the currently active ACL rules in the Redis server.

pub fn acl_users<'a>() -> Cmd

Shows a list of all the usernames of the currently configured users in the Redis ACL system.

pub fn acl_getuser<K, 'a>(username: K) -> Cmd
where K: ToRedisArgs,

Returns all the rules defined for an existing ACL user.

pub fn acl_setuser<K, 'a>(username: K) -> Cmd
where K: ToRedisArgs,

Creates an ACL user without any privilege.

pub fn acl_setuser_rules<K, 'a>(username: K, rules: &'a [Rule]) -> Cmd
where K: ToRedisArgs,

Creates an ACL user with the specified rules or modify the rules of an existing user.

pub fn acl_deluser<K, 'a>(usernames: &'a [K]) -> Cmd
where K: ToRedisArgs,

Delete all the specified ACL users and terminate all the connections that are authenticated with such users.

pub fn acl_cat<'a>() -> Cmd

Shows the available ACL categories.

pub fn acl_cat_categoryname<K, 'a>(categoryname: K) -> Cmd
where K: ToRedisArgs,

Shows all the Redis commands in the specified category.

pub fn acl_genpass<'a>() -> Cmd

Generates a 256-bits password starting from /dev/urandom if available.

pub fn acl_genpass_bits<'a>(bits: isize) -> Cmd

Generates a 1-to-1024-bits password starting from /dev/urandom if available.

pub fn acl_whoami<'a>() -> Cmd

Returns the username the current connection is authenticated with.

pub fn acl_log<'a>(count: isize) -> Cmd

Shows a list of recent ACL security events

pub fn acl_log_reset<'a>() -> Cmd

Clears the ACL log.

pub fn acl_help<'a>() -> Cmd

Returns a helpful text describing the different subcommands.

pub fn geo_add<K, M, 'a>(key: K, members: M) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs,

Adds the specified geospatial items to the specified key.

Every member has to be written as a tuple of (longitude, latitude, member_name). It can be a single tuple, or a vector of tuples.

longitude, latitude can be set using redis::geo::Coord.

Returns the number of elements added to the sorted set, not including elements already existing for which the score was updated.

Example

use redis::{Commands, Connection, RedisResult};
use redis::geo::Coord;

fn add_point(con: &mut Connection) -> RedisResult<isize> {
    con.geo_add("my_gis", (Coord::lon_lat(13.361389, 38.115556), "Palermo"))
}

fn add_point_with_tuples(con: &mut Connection) -> RedisResult<isize> {
    con.geo_add("my_gis", ("13.361389", "38.115556", "Palermo"))
}

fn add_many_points(con: &mut Connection) -> RedisResult<isize> {
    con.geo_add("my_gis", &[
        ("13.361389", "38.115556", "Palermo"),
        ("15.087269", "37.502669", "Catania")
    ])
}

pub fn geo_dist<K, M1, M2, 'a>( key: K, member1: M1, member2: M2, unit: Unit ) -> Cmd
where K: ToRedisArgs, M1: ToRedisArgs, M2: ToRedisArgs,

Return the distance between two members in the geospatial index represented by the sorted set.

If one or both the members are missing, the command returns NULL, so it may be convenient to parse its response as either Option<f64> or Option<String>.

Example

use redis::{Commands, RedisResult};
use redis::geo::Unit;

fn get_dists(con: &mut redis::Connection) {
    let x: RedisResult<f64> = con.geo_dist(
        "my_gis",
        "Palermo",
        "Catania",
        Unit::Kilometers
    );
    // x is Ok(166.2742)

    let x: RedisResult<Option<f64>> = con.geo_dist(
        "my_gis",
        "Palermo",
        "Atlantis",
        Unit::Meters
    );
    // x is Ok(None)
}

pub fn geo_hash<K, M, 'a>(key: K, members: M) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs,

Return valid Geohash strings representing the position of one or more members of the geospatial index represented by the sorted set at key.

Example

use redis::{Commands, RedisResult};

fn get_hash(con: &mut redis::Connection) {
    let x: RedisResult<Vec<String>> = con.geo_hash("my_gis", "Palermo");
    // x is vec!["sqc8b49rny0"]

    let x: RedisResult<Vec<String>> = con.geo_hash("my_gis", &["Palermo", "Catania"]);
    // x is vec!["sqc8b49rny0", "sqdtr74hyu0"]
}

pub fn geo_pos<K, M, 'a>(key: K, members: M) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs,

Return the positions of all the specified members of the geospatial index represented by the sorted set at key.

Every position is a pair of (longitude, latitude). redis::geo::Coord can be used to convert these value in a struct.

Example

use redis::{Commands, RedisResult};
use redis::geo::Coord;

fn get_position(con: &mut redis::Connection) {
    let x: RedisResult<Vec<Vec<f64>>> = con.geo_pos("my_gis", &["Palermo", "Catania"]);
    // x is [ [ 13.361389, 38.115556 ], [ 15.087269, 37.502669 ] ];

    let x: Vec<Coord<f64>> = con.geo_pos("my_gis", "Palermo").unwrap();
    // x[0].longitude is 13.361389
    // x[0].latitude is 38.115556
}

pub fn geo_radius<K, 'a>( key: K, longitude: f64, latitude: f64, radius: f64, unit: Unit, options: RadiusOptions ) -> Cmd
where K: ToRedisArgs,

Return the members of a sorted set populated with geospatial information using geo_add, which are within the borders of the area specified with the center location and the maximum distance from the center (the radius).

Every item in the result can be read with redis::geo::RadiusSearchResult, which support the multiple formats returned by GEORADIUS.

use redis::{Commands, RedisResult};
use redis::geo::{RadiusOptions, RadiusSearchResult, RadiusOrder, Unit};

fn radius(con: &mut redis::Connection) -> Vec<RadiusSearchResult> {
    let opts = RadiusOptions::default().with_dist().order(RadiusOrder::Asc);
    con.geo_radius("my_gis", 15.90, 37.21, 51.39, Unit::Kilometers, opts).unwrap()
}

pub fn geo_radius_by_member<K, M, 'a>( key: K, member: M, radius: f64, unit: Unit, options: RadiusOptions ) -> Cmd
where K: ToRedisArgs, M: ToRedisArgs,

Retrieve members selected by distance with the center of member. The member itself is always contained in the results.

pub fn xack<K, G, I, 'a>(key: K, group: G, ids: &'a [I]) -> Cmd
where K: ToRedisArgs, G: ToRedisArgs, I: ToRedisArgs,

Ack pending stream messages checked out by a consumer.

XACK <key> <group> <id> <id> ... <id>

pub fn xadd<K, ID, F, V, 'a>(key: K, id: ID, items: &'a [(F, V)]) -> Cmd
where K: ToRedisArgs, ID: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs,

Add a stream message by key. Use * as the id for the current timestamp.

XADD key <ID or *> [field value] [field value] ...

pub fn xadd_map<K, ID, BTM, 'a>(key: K, id: ID, map: BTM) -> Cmd
where K: ToRedisArgs, ID: ToRedisArgs, BTM: ToRedisArgs,

BTreeMap variant for adding a stream message by key. Use * as the id for the current timestamp.

XADD key <ID or *> [rust BTreeMap] ...

pub fn xadd_maxlen<K, ID, F, V, 'a>( key: K, maxlen: StreamMaxlen, id: ID, items: &'a [(F, V)] ) -> Cmd
where K: ToRedisArgs, ID: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs,

Add a stream message while capping the stream at a maxlength.

XADD key [MAXLEN [~|=] <count>] <ID or *> [field value] [field value] ...

pub fn xadd_maxlen_map<K, ID, BTM, 'a>( key: K, maxlen: StreamMaxlen, id: ID, map: BTM ) -> Cmd
where K: ToRedisArgs, ID: ToRedisArgs, BTM: ToRedisArgs,

BTreeMap variant for adding a stream message while capping the stream at a maxlength.

XADD key [MAXLEN [~|=] <count>] <ID or *> [rust BTreeMap] ...

pub fn xclaim<K, G, C, MIT, ID, 'a>( key: K, group: G, consumer: C, min_idle_time: MIT, ids: &'a [ID] ) -> Cmd
where K: ToRedisArgs, G: ToRedisArgs, C: ToRedisArgs, MIT: ToRedisArgs, ID: ToRedisArgs,

Claim pending, unacked messages, after some period of time, currently checked out by another consumer.

This method only accepts the must-have arguments for claiming messages. If optional arguments are required, see xclaim_options below.

XCLAIM <key> <group> <consumer> <min-idle-time> [<ID-1> <ID-2>]

pub fn xclaim_options<K, G, C, MIT, ID, 'a>( key: K, group: G, consumer: C, min_idle_time: MIT, ids: &'a [ID], options: StreamClaimOptions ) -> Cmd
where K: ToRedisArgs, G: ToRedisArgs, C: ToRedisArgs, MIT: ToRedisArgs, ID: ToRedisArgs,

This is the optional arguments version for claiming unacked, pending messages currently checked out by another consumer.

use redis::{Connection,Commands,RedisResult};
use redis::streams::{StreamClaimOptions,StreamClaimReply};
let client = redis::Client::open("redis://127.0.0.1/0").unwrap();
let mut con = client.get_connection().unwrap();

// Claim all pending messages for key "k1",
// from group "g1", checked out by consumer "c1"
// for 10ms with RETRYCOUNT 2 and FORCE

let opts = StreamClaimOptions::default()
    .with_force()
    .retry(2);
let results: RedisResult<StreamClaimReply> =
    con.xclaim_options("k1", "g1", "c1", 10, &["0"], opts);

// All optional arguments return a `Result<StreamClaimReply>` with one exception:
// Passing JUSTID returns only the message `id` and omits the HashMap for each message.

let opts = StreamClaimOptions::default()
    .with_justid();
let results: RedisResult<Vec<String>> =
    con.xclaim_options("k1", "g1", "c1", 10, &["0"], opts);

XCLAIM <key> <group> <consumer> <min-idle-time> <ID-1> <ID-2>
    [IDLE <milliseconds>] [TIME <mstime>] [RETRYCOUNT <count>]
    [FORCE] [JUSTID]

pub fn xdel<K, ID, 'a>(key: K, ids: &'a [ID]) -> Cmd
where K: ToRedisArgs, ID: ToRedisArgs,

Deletes a list of ids for a given stream key.

XDEL <key> [<ID1> <ID2> ... <IDN>]

pub fn xgroup_create<K, G, ID, 'a>(key: K, group: G, id: ID) -> Cmd
where K: ToRedisArgs, G: ToRedisArgs, ID: ToRedisArgs,

This command is used for creating a consumer group. It expects the stream key to already exist. Otherwise, use xgroup_create_mkstream if it doesn’t. The id is the starting message id all consumers should read from. Use $ If you want all consumers to read from the last message added to stream.

XGROUP CREATE <key> <groupname> <id or $>

pub fn xgroup_create_mkstream<K, G, ID, 'a>(key: K, group: G, id: ID) -> Cmd
where K: ToRedisArgs, G: ToRedisArgs, ID: ToRedisArgs,

This is the alternate version for creating a consumer group which makes the stream if it doesn’t exist.

XGROUP CREATE <key> <groupname> <id or $> [MKSTREAM]

pub fn xgroup_setid<K, G, ID, 'a>(key: K, group: G, id: ID) -> Cmd
where K: ToRedisArgs, G: ToRedisArgs, ID: ToRedisArgs,

Alter which id you want consumers to begin reading from an existing consumer group.

XGROUP SETID <key> <groupname> <id or $>

pub fn xgroup_destroy<K, G, 'a>(key: K, group: G) -> Cmd
where K: ToRedisArgs, G: ToRedisArgs,

Destroy an existing consumer group for a given stream key

XGROUP SETID <key> <groupname> <id or $>

pub fn xgroup_delconsumer<K, G, C, 'a>(key: K, group: G, consumer: C) -> Cmd
where K: ToRedisArgs, G: ToRedisArgs, C: ToRedisArgs,

This deletes a consumer from an existing consumer group for given stream `key.

XGROUP DELCONSUMER <key> <groupname> <consumername>

pub fn xinfo_consumers<K, G, 'a>(key: K, group: G) -> Cmd
where K: ToRedisArgs, G: ToRedisArgs,

This returns all info details about which consumers have read messages for given consumer group. Take note of the StreamInfoConsumersReply return type.

It’s possible this return value might not contain new fields added by Redis in future versions.

XINFO CONSUMERS <key> <group>

pub fn xinfo_groups<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Returns all consumer groups created for a given stream key. Take note of the StreamInfoGroupsReply return type.

It’s possible this return value might not contain new fields added by Redis in future versions.

XINFO GROUPS <key>

pub fn xinfo_stream<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Returns info about high-level stream details (first & last message id, length, number of groups, etc.) Take note of the StreamInfoStreamReply return type.

It’s possible this return value might not contain new fields added by Redis in future versions.

XINFO STREAM <key>

pub fn xlen<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

Returns the number of messages for a given stream key.

XLEN <key>

pub fn xpending<K, G, 'a>(key: K, group: G) -> Cmd
where K: ToRedisArgs, G: ToRedisArgs,

This is a basic version of making XPENDING command calls which only passes a stream key and consumer group and it returns details about which consumers have pending messages that haven’t been acked.

You can use this method along with xclaim or xclaim_options for determining which messages need to be retried.

Take note of the StreamPendingReply return type.

XPENDING <key> <group> [<start> <stop> <count> [<consumer>]]

pub fn xpending_count<K, G, S, E, C, 'a>( key: K, group: G, start: S, end: E, count: C ) -> Cmd
where K: ToRedisArgs, G: ToRedisArgs, S: ToRedisArgs, E: ToRedisArgs, C: ToRedisArgs,

This XPENDING version returns a list of all messages over the range. You can use this for paginating pending messages (but without the message HashMap).

Start and end follow the same rules xrange args. Set start to - and end to + for the entire stream.

Take note of the StreamPendingCountReply return type.

XPENDING <key> <group> <start> <stop> <count>

pub fn xpending_consumer_count<K, G, S, E, C, CN, 'a>( key: K, group: G, start: S, end: E, count: C, consumer: CN ) -> Cmd
where K: ToRedisArgs, G: ToRedisArgs, S: ToRedisArgs, E: ToRedisArgs, C: ToRedisArgs, CN: ToRedisArgs,

An alternate version of xpending_count which filters by consumer name.

Start and end follow the same rules xrange args. Set start to - and end to + for the entire stream.

Take note of the StreamPendingCountReply return type.

XPENDING <key> <group> <start> <stop> <count> <consumer>

pub fn xrange<K, S, E, 'a>(key: K, start: S, end: E) -> Cmd
where K: ToRedisArgs, S: ToRedisArgs, E: ToRedisArgs,

Returns a range of messages in a given stream key.

Set start to - to begin at the first message. Set end to + to end the most recent message. You can pass message id to both start and end.

Take note of the StreamRangeReply return type.

XRANGE key start end

pub fn xrange_all<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

A helper method for automatically returning all messages in a stream by key. Use with caution!

XRANGE key - +

pub fn xrange_count<K, S, E, C, 'a>(key: K, start: S, end: E, count: C) -> Cmd
where K: ToRedisArgs, S: ToRedisArgs, E: ToRedisArgs, C: ToRedisArgs,

A method for paginating a stream by key.

XRANGE key start end [COUNT <n>]

pub fn xread<K, ID, 'a>(keys: &'a [K], ids: &'a [ID]) -> Cmd
where K: ToRedisArgs, ID: ToRedisArgs,

Read a list of ids for each stream key. This is the basic form of reading streams. For more advanced control, like blocking, limiting, or reading by consumer group, see xread_options.

XREAD STREAMS key_1 key_2 ... key_N ID_1 ID_2 ... ID_N

pub fn xread_options<K, ID, 'a>( keys: &'a [K], ids: &'a [ID], options: &'a StreamReadOptions ) -> Cmd
where K: ToRedisArgs, ID: ToRedisArgs,

This method handles setting optional arguments for XREAD or XREADGROUP Redis commands.

use redis::{Connection,RedisResult,Commands};
use redis::streams::{StreamReadOptions,StreamReadReply};
let client = redis::Client::open("redis://127.0.0.1/0").unwrap();
let mut con = client.get_connection().unwrap();

// Read 10 messages from the start of the stream,
// without registering as a consumer group.

let opts = StreamReadOptions::default()
    .count(10);
let results: RedisResult<StreamReadReply> =
    con.xread_options(&["k1"], &["0"], &opts);

// Read all undelivered messages for a given
// consumer group. Be advised: the consumer group must already
// exist before making this call. Also note: we're passing
// '>' as the id here, which means all undelivered messages.

let opts = StreamReadOptions::default()
    .group("group-1", "consumer-1");
let results: RedisResult<StreamReadReply> =
    con.xread_options(&["k1"], &[">"], &opts);

XREAD [BLOCK <milliseconds>] [COUNT <count>]
    STREAMS key_1 key_2 ... key_N
    ID_1 ID_2 ... ID_N

XREADGROUP [BLOCK <milliseconds>] [COUNT <count>] [NOACK] [GROUP group-name consumer-name]
    STREAMS key_1 key_2 ... key_N
    ID_1 ID_2 ... ID_N

pub fn xrevrange<K, E, S, 'a>(key: K, end: E, start: S) -> Cmd
where K: ToRedisArgs, E: ToRedisArgs, S: ToRedisArgs,

This is the reverse version of xrange. The same rules apply for start and end here.

XREVRANGE key end start

pub fn xrevrange_all<K, 'a>(key: K) -> Cmd
where K: ToRedisArgs,

This is the reverse version of xrange_all. The same rules apply for start and end here.

XREVRANGE key + -

pub fn xrevrange_count<K, E, S, C, 'a>( key: K, end: E, start: S, count: C ) -> Cmd
where K: ToRedisArgs, E: ToRedisArgs, S: ToRedisArgs, C: ToRedisArgs,

This is the reverse version of xrange_count. The same rules apply for start and end here.

XREVRANGE key end start [COUNT <n>]

pub fn xtrim<K, 'a>(key: K, maxlen: StreamMaxlen) -> Cmd
where K: ToRedisArgs,

Trim a stream key to a MAXLEN count.

XTRIM <key> MAXLEN [~|=] <count>  (Same as XADD MAXLEN option)