Crate mod_redis

Modules

test

Structs

ClientManager

Cmd

Represents redis commands.

RedisConnKey

RedisConnection

RedisError

Represents a redis error.

Script

Represents a lua script.

ScriptInvocation

Represents a prepared script call.

Enums

ClientWrapper

ConnectionWrapper

NodeSpec

RedisValue

Internal low-level redis value enum.

Traits

FromRedisValue

This trait is used to convert a redis value into a more appropriate type.

Functions

build_cmd

cmd

Shortcut function to creating a command with a single argument.

register